functionalities=>{updatePost, logOut, hosting}

+++ js/navbar.js
@@ -10,19 +10,42 @@ document
 document
   .querySelector(".toggle")
   .addEventListener("click", handleShowLinkMobile);
+
+// logout button
+
 const logout = document.getElementById("logout");
-// const logOptions = document.querySelector(".log-options");
-// if (localStorage.key("token")) {
-//   let login;
-//   login = `<a href="#">Login</a>`;
-//   logOptions.innerHTML = login;
-//   console.log(logOptions.innerHTML);
-//   location.href = "http://127.0.0.1:5500/index.html";
-// }
-logout.addEventListener("click", (e) => {
-  e.preventDefault();
-  // let login;
-  // login = `<a href="#">Logout</a>`;
-  // logOptions.innerHTML = login;
-  localStorage.clear("token");
-});
+const changeButton = document.getElementById("logout");
+if (localStorage.getItem("token") === null) {
+  let value = "";
+  value += `<a href="#">Login</a>`;
+  changeButton.innerHTML = value;
+  logout.addEventListener("click", (e) => {
+    e.preventDefault();
+    window.location.href = "./login.html";
+  });
+} else if (localStorage.getItem("token")) {
+  let value = "";
+  value += `<a href="#">Logout</a>`;
+  changeButton.innerHTML = value;
+  logout.addEventListener("click", (e) => {
+    e.preventDefault();
+
+    localStorage.clear("token");
+    window.location.href = "./index.html";
+  });
+}
+
+// create button
+
+const create = document.getElementById("create");
+if (localStorage.getItem("token") === null) {
+  create.addEventListener("click", (e) => {
+    e.preventDefault();
+    window.location.href = "./login.html";
+  });
+} else if (localStorage.getItem("token")) {
+  create.addEventListener("click", (e) => {
+    e.preventDefault();
+    window.location.href = "./create.html";
+  });
+}
